second commit

+++ src/app/pages/events/list/events.component.ts
@@ -1,6 +1,9 @@
-import {Component, OnInit} from '@angular/core';
+import {Component, OnInit, ViewContainerRef} from '@angular/core';
 import {Router} from '@angular/router';
 import {EventService} from '../../../service/EventService';
+import {forEach} from '@angular/router/src/utils/collection';
+import {DatePipe} from '@angular/common';
+import {ToastsManager} from 'ng2-toastr';
 
 
 @Component({
@@ -11,14 +14,18 @@ import {EventService} from '../../../service/EventService';
 export class EventsComponent implements OnInit {
   searchText
   events;
+  event;
 
-  constructor(private router: Router, public eventService: EventService) {
+  constructor(public toastr: ToastsManager, vcr: ViewContainerRef, private router: Router, public eventService: EventService) {
+    this.toastr.setRootViewContainerRef(vcr);
   }
 
   ngOnInit() {
     this.eventService.getEvents()
       .subscribe(data => {
           this.events = data;
+          // this.events.forEach(e => {
+          // })
         },
         err => {
           console.log(err);
@@ -38,10 +45,22 @@ export class EventsComponent implements OnInit {
       .subscribe(data => {
         this.events.splice(this.events.indexOf(events), 1);
       }, err => {
+        this.toastr.success('event ' + events.title + ' deleted', 'Success!');
         console.log(err);
       });
-    this.eventService.deleteEventFromAccpetedUsers(events.id)
+    this.eventService.getEvent(events.id)
       .subscribe(data => {
+        this.event = data;
+        if (this.event != null) {
+          this.eventService.deleteEventFromAccpetedUsers(events.id)
+            .subscribe(data1 => {
+              this.toastr.success('event:' + events.title + ' deleted', 'Success!');
+            }, err => {
+              console.log(err);
+            });
+        } else {
+          this.toastr.success('event: ' + events.title + ' deleted', 'Success!');
+        }
       }, err => {
         console.log(err);
       });
